requirements.txt
streamlit
streamlit-folium
folium
pandas
geopy
pyrebase4
gcloud==0.18.3